promote LockAndTimestampServices
