+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/LockAndTimestampServiceFactory.java
@@ -34,7 +34,6 @@ import com.palantir.atlasdb.config.TimeLockRequestBatcherProviders;
 import com.palantir.atlasdb.debug.LockDiagnosticComponents;
 import com.palantir.atlasdb.debug.LockDiagnosticConjureTimelockService;
 import com.palantir.atlasdb.factory.Leaders.LocalPaxosServices;
-import com.palantir.atlasdb.factory.TransactionManagers.LockAndTimestampServices;
 import com.palantir.atlasdb.factory.startup.TimeLockMigrator;
 import com.palantir.atlasdb.factory.timelock.TimestampCorroboratingTimelockService;
 import com.palantir.atlasdb.http.AtlasDbHttpClients;
