+++ atlasdb-config/src/test/java/com/palantir/atlasdb/factory/TransactionManagersTest.java
@@ -156,8 +156,7 @@ public class TransactionManagersTest {
             SslConfiguration.of(Paths.get("var/security/trustStore.jks"));
 
     private final TimeLockMigrator migrator = mock(TimeLockMigrator.class);
-    private final TransactionManagers.LockAndTimestampServices lockAndTimestampServices =
-            mock(TransactionManagers.LockAndTimestampServices.class);
+    private final LockAndTimestampServices lockAndTimestampServices = mock(LockAndTimestampServices.class);
     private final MetricsManager metricsManager = MetricsManagers.createForTests();
     private final DialogueClients.ReloadingFactory reloadingFactory = DialogueClients.create(
             Refreshable.only(ServicesConfigBlock.builder().build()));
@@ -264,7 +263,7 @@ public class TransactionManagersTest {
         setUpForRemoteServices();
         setUpLeaderBlockInConfig();
 
-        TransactionManagers.LockAndTimestampServices lockAndTimestamp = getLockAndTimestampServices();
+        LockAndTimestampServices lockAndTimestamp = getLockAndTimestampServices();
         availableServer.verify(getRequestedFor(urlMatching(LEADER_UUID_PATH)));
 
         lockAndTimestamp.timelock().getFreshTimestamp();
@@ -282,7 +281,7 @@ public class TransactionManagersTest {
         setUpForLocalServices();
         setUpLeaderBlockInConfig();
 
-        TransactionManagers.LockAndTimestampServices lockAndTimestamp = getLockAndTimestampServices();
+        LockAndTimestampServices lockAndTimestamp = getLockAndTimestampServices();
         availableServer.verify(getRequestedFor(urlMatching(LEADER_UUID_PATH)));
 
         lockAndTimestamp.timelock().getFreshTimestamp();
@@ -324,21 +323,20 @@ public class TransactionManagersTest {
         };
 
         InMemoryTimestampService ts = new InMemoryTimestampService();
-        TransactionManagers.LockAndTimestampServices lockAndTimestamp =
-                LockAndTimestampServiceFactory.createLockAndTimestampServices(
-                        metricsManager,
-                        config,
-                        Refreshable.only(mockAtlasDbRuntimeConfig),
-                        environment,
-                        lockServiceSupplier,
-                        () -> ts,
-                        invalidator,
-                        USER_AGENT,
-                        Optional.empty(),
-                        reloadingFactory,
-                        Optional.empty(),
-                        Optional.empty(),
-                        ImmutableSet.of());
+        LockAndTimestampServices lockAndTimestamp = LockAndTimestampServiceFactory.createLockAndTimestampServices(
+                metricsManager,
+                config,
+                Refreshable.only(mockAtlasDbRuntimeConfig),
+                environment,
+                lockServiceSupplier,
+                () -> ts,
+                invalidator,
+                USER_AGENT,
+                Optional.empty(),
+                reloadingFactory,
+                Optional.empty(),
+                Optional.empty(),
+                ImmutableSet.of());
 
         LockRequest lockRequest = LockRequest.builder(
                         ImmutableSortedMap.of(StringLockDescriptor.of("foo"), LockMode.WRITE))
@@ -791,7 +789,7 @@ public class TransactionManagersTest {
                 .isEqualTo(0L);
         assertThat(metricsManager.getRegistry().timer(lockMetric).getCount()).isEqualTo(0L);
 
-        TransactionManagers.LockAndTimestampServices lockAndTimestamp = getLockAndTimestampServices();
+        LockAndTimestampServices lockAndTimestamp = getLockAndTimestampServices();
         lockAndTimestamp.timelock().getFreshTimestamp();
         lockAndTimestamp.timelock().currentTimeMillis();
 
@@ -814,7 +812,7 @@ public class TransactionManagersTest {
         assertThat(metricsManager.getTaggedRegistry().timer(lockMetricName).getCount())
                 .isEqualTo(0L);
 
-        TransactionManagers.LockAndTimestampServices lockAndTimestamp = getLockAndTimestampServices();
+        LockAndTimestampServices lockAndTimestamp = getLockAndTimestampServices();
         lockAndTimestamp.timelock().getFreshTimestamp();
         lockAndTimestamp.timelock().currentTimeMillis();
 
@@ -870,7 +868,7 @@ public class TransactionManagersTest {
                         .build()));
     }
 
-    private TransactionManagers.LockAndTimestampServices getLockAndTimestampServices() {
+    private LockAndTimestampServices getLockAndTimestampServices() {
         InMemoryTimestampService ts = new InMemoryTimestampService();
         return LockAndTimestampServiceFactory.createLockAndTimestampServices(
                 metricsManager,
@@ -894,21 +892,20 @@ public class TransactionManagersTest {
 
     private void verifyUserAgentOnTimestampAndLockRequests(String timestampPath, String lockPath) {
         InMemoryTimestampService ts = new InMemoryTimestampService();
-        TransactionManagers.LockAndTimestampServices lockAndTimestamp =
-                LockAndTimestampServiceFactory.createLockAndTimestampServices(
-                        metricsManager,
-                        config,
-                        Refreshable.only(mockAtlasDbRuntimeConfig),
-                        environment,
-                        LockServiceImpl::create,
-                        () -> ts,
-                        invalidator,
-                        USER_AGENT,
-                        Optional.empty(),
-                        reloadingFactory,
-                        Optional.empty(),
-                        Optional.empty(),
-                        ImmutableSet.of());
+        LockAndTimestampServices lockAndTimestamp = LockAndTimestampServiceFactory.createLockAndTimestampServices(
+                metricsManager,
+                config,
+                Refreshable.only(mockAtlasDbRuntimeConfig),
+                environment,
+                LockServiceImpl::create,
+                () -> ts,
+                invalidator,
+                USER_AGENT,
+                Optional.empty(),
+                reloadingFactory,
+                Optional.empty(),
+                Optional.empty(),
+                ImmutableSet.of());
 
         lockAndTimestamp.timelock().getFreshTimestamp();
         lockAndTimestamp.timelock().currentTimeMillis();
