+++ atlasdb-dagger/src/main/java/com/palantir/atlasdb/services/test/TestTransactionManagerModule.java
@@ -24,7 +24,7 @@ import com.palantir.atlasdb.cleaner.Follower;
 import com.palantir.atlasdb.cleaner.api.Cleaner;
 import com.palantir.atlasdb.config.AtlasDbConfig;
 import com.palantir.atlasdb.debug.ConflictTracer;
-import com.palantir.atlasdb.factory.TransactionManagers;
+import com.palantir.atlasdb.factory.LockAndTimestampServices;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.services.ServicesConfig;
 import com.palantir.atlasdb.sweep.queue.MultiTableSweepQueueWriter;
@@ -95,7 +95,7 @@ public class TestTransactionManagerModule {
             MetricsManager metricsManager,
             ServicesConfig config,
             @Named("kvs") KeyValueService kvs,
-            TransactionManagers.LockAndTimestampServices lts,
+            LockAndTimestampServices lts,
             LockClient lockClient,
             TransactionService transactionService,
             ConflictDetectionManager conflictManager,
